v0.17 - Hybrid Cosine and Pearson Similarity
I created a new class CosinePearsonHybrid that averages the cosine
similarity with pearson.

I also tried some weighted averages and a section of  my results are
below.

...
RMSE: 0.9862317429779974  user Cosine w: 2  user Pearson w:8 item Cosine
w:2  item Pearson w:8
RMSE: 0.9848458321532898  user Cosine w: 2  user Pearson w:8 item Cosine
w:3  item Pearson w:7
RMSE: 0.9838114250082263  user Cosine w: 2  user Pearson w:8 item Cosine
w:4  item Pearson w:6
RMSE: 0.9832487804280765  user Cosine w: 2  user Pearson w:8 item Cosine
w:5  item Pearson w:5
RMSE: 0.9831332928989266  user Cosine w: 2  user Pearson w:8 item Cosine
w:6  item Pearson w:4
RMSE: 0.98343696653626    user Cosine w: 2  user Pearson w:8 item Cosine
w:7  item Pearson w:3
> RMSE: 0.9829833320917658  user Cosine w: 2  user Pearson w:8 item
Cosine w:8  item Pearson w:2
RMSE: 0.9835743123981594  user Cosine w: 2  user Pearson w:8 item Cosine
w:9  item Pearson w:1
RMSE: 0.9885548189401521  user Cosine w: 3  user Pearson w:7 item Cosine
w:1  item Pearson w:9
RMSE: 0.9859485398008404  user Cosine w: 3  user Pearson w:7 item Cosine
w:2  item Pearson w:8
RMSE: 0.9846814210724572  user Cosine w: 3  user Pearson w:7 item Cosine
w:3  item Pearson w:7
...

The best combination of weighted averages was  

User Based CF -  Cosine Weight :  2     Pearson Weight:  8
Item Base CF -   Cosine Weight :  8     Pearson Weight:  2

RMSE: 0.9829833320917658

+++ src/similarity/metric/Cosine.java
@@ -45,7 +45,7 @@ public class Cosine implements SimilarityMetric
 		double n2 = p2.getNorm();
 		
 		// (Math.min(Math.abs((double)common.size()), max)/max  < significance weighting 	
-		return (n1 > 0 && n2 > 0) ? (dotProduct / (n1 * n2)) * (Math.min(Math.abs((double)common.size()), max)/max )  : 0;
+		return (n1 > 0 && n2 > 0) ? (dotProduct / (n1 * n2)) * (Math.min(Math.abs(common.size()), max)/max )  : 0;
 	}
